test: Noteのテストを分離
